assignment submission
still not super comfortable with github. i've been working on this code in the repl.it link provided for a few days, and copied/pasted my code from there into here. submitting this a little early in case something is wrong with my submission, i'll have time to correct it before the due date.